refactor(settings-overlay): extract reset section
Extract reset confirmation UI and behavior into ResetSettingsSection component
and useResetConfirmation hook. Preserves two-click confirmation, 3-second
auto-cancel, dispatchReset(), synth.setMute(false), and close-on-confirm
behavior.
